airflow-orchestration/pyproject.toml
[project]
name = "airflow-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "airflow-provider-duckdb>=0.2.0",
    "apache-airflow==2.11.0",
    "apache-airflow-providers-common-sql>=1.27.1",
    "apache-airflow-providers-postgres>=6.2.0",
    "apache-airflow-providers-standard>=1.2.0",
    "astronomer-cosmos>=1.10.1",
    "deltalake>=1.0.2",
    "pandas>=2.3.0",
    "psycopg[binary]>=3.2.9",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-sugar>=1.0.0",
    "ruff>=0.11.2",
]
